Move layout files to layouts folder
